Move AD test files up a directory
