+++ x-pack/test/functional/apps/ml/group2/advanced_job.ts
@@ -5,7 +5,7 @@
  * 2.0.
  */
 
-import { FtrProviderContext } from '../../../../ftr_provider_context';
+import { FtrProviderContext } from '../../../ftr_provider_context';
 
 interface Detector {
   identifier: string;
