+++ x-pack/test/functional/apps/ml/group2/annotations.ts
@@ -6,7 +6,7 @@
  */
 
 import { Annotation } from '@kbn/ml-plugin/common/types/annotations';
-import { FtrProviderContext } from '../../../../ftr_provider_context';
+import { FtrProviderContext } from '../../../ftr_provider_context';
 
 export default function ({ getService }: FtrProviderContext) {
   const esArchiver = getService('esArchiver');
