+++ x-pack/test/functional/apps/ml/group2/custom_urls.ts
@@ -7,12 +7,12 @@
 
 import { Job, Datafeed } from '@kbn/ml-plugin/common/types/anomaly_detection_jobs';
 import { TIME_RANGE_TYPE } from '@kbn/ml-plugin/public/application/jobs/components/custom_url_editor/constants';
-import { FtrProviderContext } from '../../../../ftr_provider_context';
+import { FtrProviderContext } from '../../../ftr_provider_context';
 import type {
   DiscoverUrlConfig,
   DashboardUrlConfig,
   OtherUrlConfig,
-} from '../../../../services/ml/job_table';
+} from '../../../services/ml/job_table';
 
 // @ts-expect-error doesn't implement the full interface
 const JOB_CONFIG: Job = {
