+++ x-pack/test/functional/apps/ml/group2/index.ts
@@ -11,7 +11,9 @@ export default function ({ getService, loadTestFile }: FtrProviderContext) {
   const esArchiver = getService('esArchiver');
   const ml = getService('ml');
 
-  describe('machine learning - group 2', () => {
+  describe('machine learning - anomaly detection', function () {
+    this.tags(['skipFirefox']);
+
     before(async () => {
       await ml.securityCommon.createMlRoles();
       await ml.securityCommon.createMlUsers();
@@ -37,6 +39,19 @@ export default function ({ getService, loadTestFile }: FtrProviderContext) {
       await ml.testResources.resetKibanaTimeZone();
     });
 
-    loadTestFile(require.resolve('./anomaly_detection'));
+    loadTestFile(require.resolve('./single_metric_job'));
+    loadTestFile(require.resolve('./single_metric_job_without_datafeed_start'));
+    loadTestFile(require.resolve('./multi_metric_job'));
+    loadTestFile(require.resolve('./population_job'));
+    loadTestFile(require.resolve('./saved_search_job'));
+    loadTestFile(require.resolve('./advanced_job'));
+    loadTestFile(require.resolve('./single_metric_viewer'));
+    loadTestFile(require.resolve('./anomaly_explorer'));
+    loadTestFile(require.resolve('./categorization_job'));
+    loadTestFile(require.resolve('./date_nanos_job'));
+    loadTestFile(require.resolve('./annotations'));
+    loadTestFile(require.resolve('./aggregated_scripted_job'));
+    loadTestFile(require.resolve('./custom_urls'));
+    loadTestFile(require.resolve('./forecasts'));
   });
 }
