+++ x-pack/test/functional/apps/ml/group2/multi_metric_job.ts
@@ -5,67 +5,53 @@
  * 2.0.
  */
 
-import { FtrProviderContext } from '../../../../ftr_provider_context';
+import { FtrProviderContext } from '../../../ftr_provider_context';
 
 export default function ({ getService }: FtrProviderContext) {
   const esArchiver = getService('esArchiver');
   const ml = getService('ml');
 
-  const jobId = `ec_population_1_${Date.now()}`;
+  const jobId = `fq_multi_1_${Date.now()}`;
   const jobIdClone = `${jobId}_clone`;
   const jobDescription =
-    'Create population job based on the ecommerce sample dataset with 2h bucketspan over customer_id' +
-    ' - detectors: (Mean(products.base_price) by customer_gender), (Mean(products.quantity) by category.leyword)';
-  const jobGroups = ['automated', 'ecommerce', 'population'];
+    'Create multi metric job based on the farequote dataset with 15m bucketspan and min/max/mean(responsetime) split by airline';
+  const jobGroups = ['automated', 'farequote', 'multi-metric'];
   const jobGroupsClone = [...jobGroups, 'clone'];
-  const populationField = 'customer_id';
-  const detectors = [
-    {
-      identifier: 'Mean(products.base_price)',
-      splitField: 'customer_gender',
-      frontCardTitle: 'FEMALE',
-      numberOfBackCards: 1,
-    },
-    {
-      identifier: 'Mean(products.quantity)',
-      splitField: 'category.keyword',
-      frontCardTitle: "Men's Clothing",
-      numberOfBackCards: 5,
-    },
-  ];
-  const bucketSpan = '2h';
-  const memoryLimit = '8mb';
+  const aggAndFieldIdentifiers = ['Min(responsetime)', 'Max(responsetime)', 'Mean(responsetime)'];
+  const splitField = 'airline';
+  const bucketSpan = '15m';
+  const memoryLimit = '20mb';
 
   function getExpectedRow(expectedJobId: string, expectedJobGroups: string[]) {
     return {
       id: expectedJobId,
       description: jobDescription,
       jobGroups: [...new Set(expectedJobGroups)].sort(),
-      recordCount: '4,675',
+      recordCount: '86,274',
       memoryStatus: 'ok',
       jobState: 'closed',
       datafeedState: 'stopped',
-      latestTimestamp: '2019-07-12 23:45:36',
+      latestTimestamp: '2016-02-11 23:59:54',
     };
   }
 
   function getExpectedCounts(expectedJobId: string) {
     return {
       job_id: expectedJobId,
-      processed_record_count: '4,675',
-      processed_field_count: '23,375',
-      input_bytes: '867.7 KB',
-      input_field_count: '23,375',
+      processed_record_count: '86,274',
+      processed_field_count: '172,548',
+      input_bytes: '6.4 MB',
+      input_field_count: '172,548',
       invalid_date_count: '0',
       missing_field_count: '0',
       out_of_order_timestamp_count: '0',
       empty_bucket_count: '0',
       sparse_bucket_count: '0',
-      bucket_count: '371',
-      earliest_record_timestamp: '2019-06-12 00:04:19',
-      latest_record_timestamp: '2019-07-12 23:45:36',
-      input_record_count: '4,675',
-      latest_bucket_timestamp: '2019-07-12 22:00:00',
+      bucket_count: '479',
+      earliest_record_timestamp: '2016-02-07 00:00:00',
+      latest_record_timestamp: '2016-02-11 23:59:54',
+      input_record_count: '86,274',
+      latest_bucket_timestamp: '2016-02-11 23:45:00',
     };
   }
 
@@ -74,22 +60,22 @@ export default function ({ getService }: FtrProviderContext) {
       job_id: expectedJobId,
       result_type: 'model_size_stats',
       model_bytes_exceeded: '0.0 B',
-      total_by_field_count: '25',
-      total_over_field_count: '92',
-      total_partition_field_count: '3',
+      total_by_field_count: '59',
+      total_over_field_count: '0',
+      total_partition_field_count: '58',
       bucket_allocation_failures_count: '0',
       memory_status: 'ok',
-      timestamp: '2019-07-12 20:00:00',
+      timestamp: '2016-02-11 23:30:00',
     };
   }
 
   const calendarId = `wizard-test-calendar_${Date.now()}`;
 
-  describe('population', function () {
+  describe('multi metric', function () {
     this.tags(['ml']);
     before(async () => {
-      await esArchiver.loadIfNeeded('x-pack/test/functional/es_archives/ml/ecommerce');
-      await ml.testResources.createIndexPatternIfNeeded('ft_ecommerce', 'order_date');
+      await esArchiver.loadIfNeeded('x-pack/test/functional/es_archives/ml/farequote');
+      await ml.testResources.createIndexPatternIfNeeded('ft_farequote', '@timestamp');
       await ml.testResources.setKibanaTimeZoneToUTC();
 
       await ml.api.createCalendar(calendarId);
@@ -98,10 +84,10 @@ export default function ({ getService }: FtrProviderContext) {
 
     after(async () => {
       await ml.api.cleanMlIndices();
-      await ml.testResources.deleteIndexPatternByTitle('ft_ecommerce');
+      await ml.testResources.deleteIndexPatternByTitle('ft_farequote');
     });
 
-    it('job creation loads the population wizard for the source data', async () => {
+    it('job creation loads the multi metric wizard for the source data', async () => {
       await ml.testExecution.logTestStep('job creation loads the job management page');
       await ml.navigation.navigateToMl();
       await ml.navigation.navigateToJobManagement();
@@ -110,20 +96,20 @@ export default function ({ getService }: FtrProviderContext) {
       await ml.jobManagement.navigateToNewJobSourceSelection();
 
       await ml.testExecution.logTestStep('job creation loads the job type selection page');
-      await ml.jobSourceSelection.selectSourceForAnomalyDetectionJob('ft_ecommerce');
+      await ml.jobSourceSelection.selectSourceForAnomalyDetectionJob('ft_farequote');
 
-      await ml.testExecution.logTestStep('job creation loads the population job wizard page');
-      await ml.jobTypeSelection.selectPopulationJob();
+      await ml.testExecution.logTestStep('job creation loads the multi metric job wizard page');
+      await ml.jobTypeSelection.selectMultiMetricJob();
     });
 
-    it('job creation navigates through the population wizard and sets all needed fields', async () => {
+    it('job creation navigates through the multi metric wizard and sets all needed fields', async () => {
       await ml.testExecution.logTestStep('job creation displays the time range step');
       await ml.jobWizardCommon.assertTimeRangeSectionExists();
 
       await ml.testExecution.logTestStep('job creation sets the time range');
       await ml.jobWizardCommon.clickUseFullDataButton(
-        'Jun 12, 2019 @ 00:04:19.000',
-        'Jul 12, 2019 @ 23:45:36.000'
+        'Feb 7, 2016 @ 00:00:00.000',
+        'Feb 11, 2016 @ 23:59:54.000'
       );
 
       await ml.testExecution.logTestStep('job creation displays the event rate chart');
@@ -133,42 +119,30 @@ export default function ({ getService }: FtrProviderContext) {
       await ml.testExecution.logTestStep('job creation displays the pick fields step');
       await ml.jobWizardCommon.advanceToPickFieldsSection();
 
-      await ml.testExecution.logTestStep('job creation selects the population field');
-      await ml.jobWizardPopulation.assertPopulationFieldInputExists();
-      await ml.jobWizardPopulation.selectPopulationField(populationField);
-
       await ml.testExecution.logTestStep(
         'job creation selects detectors and displays detector previews'
       );
-      for (const [index, detector] of detectors.entries()) {
+      for (const [index, aggAndFieldIdentifier] of aggAndFieldIdentifiers.entries()) {
         await ml.jobWizardCommon.assertAggAndFieldInputExists();
-        await ml.jobWizardCommon.selectAggAndField(detector.identifier, false);
-        await ml.jobWizardCommon.assertDetectorPreviewExists(detector.identifier, index, 'SCATTER');
+        await ml.jobWizardCommon.selectAggAndField(aggAndFieldIdentifier, false);
+        await ml.jobWizardCommon.assertDetectorPreviewExists(aggAndFieldIdentifier, index, 'LINE');
       }
 
       await ml.testExecution.logTestStep(
-        'job creation inputs detector split fields and displays split cards'
+        'job creation inputs the split field and displays split cards'
       );
-      for (const [index, detector] of detectors.entries()) {
-        await ml.jobWizardPopulation.assertDetectorSplitFieldInputExists(index);
-        await ml.jobWizardPopulation.selectDetectorSplitField(index, detector.splitField);
+      await ml.jobWizardMultiMetric.assertSplitFieldInputExists();
+      await ml.jobWizardMultiMetric.selectSplitField(splitField);
 
-        await ml.jobWizardPopulation.assertDetectorSplitExists(index);
-        await ml.jobWizardPopulation.assertDetectorSplitFrontCardTitle(
-          index,
-          detector.frontCardTitle
-        );
-        await ml.jobWizardPopulation.assertDetectorSplitNumberOfBackCards(
-          index,
-          detector.numberOfBackCards
-        );
-      }
+      await ml.jobWizardMultiMetric.assertDetectorSplitExists(splitField);
+      await ml.jobWizardMultiMetric.assertDetectorSplitFrontCardTitle('AAL');
+      await ml.jobWizardMultiMetric.assertDetectorSplitNumberOfBackCards(9);
+
+      await ml.jobWizardCommon.assertInfluencerSelection([splitField]);
 
       await ml.testExecution.logTestStep('job creation displays the influencer field');
       await ml.jobWizardCommon.assertInfluencerInputExists();
-      await ml.jobWizardCommon.assertInfluencerSelection(
-        [populationField].concat(detectors.map((detector) => detector.splitField))
-      );
+      await ml.jobWizardCommon.assertInfluencerSelection([splitField]);
 
       await ml.testExecution.logTestStep('job creation inputs the bucket span');
       await ml.jobWizardCommon.assertBucketSpanInputExists();
@@ -245,27 +219,27 @@ export default function ({ getService }: FtrProviderContext) {
       );
 
       await ml.testExecution.logTestStep('job creation has detector results');
-      for (let i = 0; i < detectors.length; i++) {
+      for (let i = 0; i < aggAndFieldIdentifiers.length; i++) {
         await ml.api.assertDetectorResultsExist(jobId, i);
       }
     });
 
-    it('job cloning opens the existing job in the population wizard', async () => {
+    it('job cloning opens the existing job in the multi metric wizard', async () => {
       await ml.testExecution.logTestStep(
-        'job cloning clicks the clone action and loads the population wizard'
+        'job cloning clicks the clone action and loads the multi metric wizard'
       );
       await ml.jobTable.clickCloneJobAction(jobId);
-      await ml.jobTypeSelection.assertPopulationJobWizardOpen();
+      await ml.jobTypeSelection.assertMultiMetricJobWizardOpen();
     });
 
-    it('job cloning navigates through the population wizard, checks and sets all needed fields', async () => {
+    it('job cloning navigates through the multi metric wizard, checks and sets all needed fields', async () => {
       await ml.testExecution.logTestStep('job cloning displays the time range step');
       await ml.jobWizardCommon.assertTimeRangeSectionExists();
 
       await ml.testExecution.logTestStep('job cloning sets the time range');
       await ml.jobWizardCommon.clickUseFullDataButton(
-        'Jun 12, 2019 @ 00:04:19.000',
-        'Jul 12, 2019 @ 23:45:36.000'
+        'Feb 7, 2016 @ 00:00:00.000',
+        'Feb 11, 2016 @ 23:59:54.000'
       );
 
       await ml.testExecution.logTestStep('job cloning displays the event rate chart');
@@ -275,35 +249,24 @@ export default function ({ getService }: FtrProviderContext) {
       await ml.testExecution.logTestStep('job cloning displays the pick fields step');
       await ml.jobWizardCommon.advanceToPickFieldsSection();
 
-      await ml.testExecution.logTestStep('job cloning pre-fills the population field');
-      await ml.jobWizardPopulation.assertPopulationFieldInputExists();
-      await ml.jobWizardPopulation.assertPopulationFieldSelection([populationField]);
-
       await ml.testExecution.logTestStep(
         'job cloning pre-fills detectors and shows preview with split cards'
       );
-      for (const [index, detector] of detectors.entries()) {
-        await ml.jobWizardCommon.assertDetectorPreviewExists(detector.identifier, index, 'SCATTER');
-
-        await ml.jobWizardPopulation.assertDetectorSplitFieldSelection(index, [
-          detector.splitField,
-        ]);
-        await ml.jobWizardPopulation.assertDetectorSplitExists(index);
-        await ml.jobWizardPopulation.assertDetectorSplitFrontCardTitle(
-          index,
-          detector.frontCardTitle
-        );
-        await ml.jobWizardPopulation.assertDetectorSplitNumberOfBackCards(
-          index,
-          detector.numberOfBackCards
-        );
+      for (const [index, aggAndFieldIdentifier] of aggAndFieldIdentifiers.entries()) {
+        await ml.jobWizardCommon.assertDetectorPreviewExists(aggAndFieldIdentifier, index, 'LINE');
       }
+
+      await ml.jobWizardMultiMetric.assertDetectorSplitExists(splitField);
+      await ml.jobWizardMultiMetric.assertDetectorSplitFrontCardTitle('AAL');
+      await ml.jobWizardMultiMetric.assertDetectorSplitNumberOfBackCards(9);
+
+      await ml.testExecution.logTestStep('job cloning pre-fills the split field');
+      await ml.jobWizardMultiMetric.assertSplitFieldInputExists();
+      await ml.jobWizardMultiMetric.assertSplitFieldSelection([splitField]);
 
       await ml.testExecution.logTestStep('job cloning pre-fills influencers');
       await ml.jobWizardCommon.assertInfluencerInputExists();
-      await ml.jobWizardCommon.assertInfluencerSelection(
-        [populationField].concat(detectors.map((detector) => detector.splitField))
-      );
+      await ml.jobWizardCommon.assertInfluencerSelection([splitField]);
 
       await ml.testExecution.logTestStep('job cloning pre-fills the bucket span');
       await ml.jobWizardCommon.assertBucketSpanInputExists();
@@ -388,7 +351,7 @@ export default function ({ getService }: FtrProviderContext) {
       );
 
       await ml.testExecution.logTestStep('job cloning has detector results');
-      for (let i = 0; i < detectors.length; i++) {
+      for (let i = 0; i < aggAndFieldIdentifiers.length; i++) {
         await ml.api.assertDetectorResultsExist(jobId, i);
       }
     });
