+++ x-pack/test/functional/apps/ml/group2/population_job.ts
@@ -5,53 +5,67 @@
  * 2.0.
  */
 
-import { FtrProviderContext } from '../../../../ftr_provider_context';
+import { FtrProviderContext } from '../../../ftr_provider_context';
 
 export default function ({ getService }: FtrProviderContext) {
   const esArchiver = getService('esArchiver');
   const ml = getService('ml');
 
-  const jobId = `fq_multi_1_${Date.now()}`;
+  const jobId = `ec_population_1_${Date.now()}`;
   const jobIdClone = `${jobId}_clone`;
   const jobDescription =
-    'Create multi metric job based on the farequote dataset with 15m bucketspan and min/max/mean(responsetime) split by airline';
-  const jobGroups = ['automated', 'farequote', 'multi-metric'];
+    'Create population job based on the ecommerce sample dataset with 2h bucketspan over customer_id' +
+    ' - detectors: (Mean(products.base_price) by customer_gender), (Mean(products.quantity) by category.leyword)';
+  const jobGroups = ['automated', 'ecommerce', 'population'];
   const jobGroupsClone = [...jobGroups, 'clone'];
-  const aggAndFieldIdentifiers = ['Min(responsetime)', 'Max(responsetime)', 'Mean(responsetime)'];
-  const splitField = 'airline';
-  const bucketSpan = '15m';
-  const memoryLimit = '20mb';
+  const populationField = 'customer_id';
+  const detectors = [
+    {
+      identifier: 'Mean(products.base_price)',
+      splitField: 'customer_gender',
+      frontCardTitle: 'FEMALE',
+      numberOfBackCards: 1,
+    },
+    {
+      identifier: 'Mean(products.quantity)',
+      splitField: 'category.keyword',
+      frontCardTitle: "Men's Clothing",
+      numberOfBackCards: 5,
+    },
+  ];
+  const bucketSpan = '2h';
+  const memoryLimit = '8mb';
 
   function getExpectedRow(expectedJobId: string, expectedJobGroups: string[]) {
     return {
       id: expectedJobId,
       description: jobDescription,
       jobGroups: [...new Set(expectedJobGroups)].sort(),
-      recordCount: '86,274',
+      recordCount: '4,675',
       memoryStatus: 'ok',
       jobState: 'closed',
       datafeedState: 'stopped',
-      latestTimestamp: '2016-02-11 23:59:54',
+      latestTimestamp: '2019-07-12 23:45:36',
     };
   }
 
   function getExpectedCounts(expectedJobId: string) {
     return {
       job_id: expectedJobId,
-      processed_record_count: '86,274',
-      processed_field_count: '172,548',
-      input_bytes: '6.4 MB',
-      input_field_count: '172,548',
+      processed_record_count: '4,675',
+      processed_field_count: '23,375',
+      input_bytes: '867.7 KB',
+      input_field_count: '23,375',
       invalid_date_count: '0',
       missing_field_count: '0',
       out_of_order_timestamp_count: '0',
       empty_bucket_count: '0',
       sparse_bucket_count: '0',
-      bucket_count: '479',
-      earliest_record_timestamp: '2016-02-07 00:00:00',
-      latest_record_timestamp: '2016-02-11 23:59:54',
-      input_record_count: '86,274',
-      latest_bucket_timestamp: '2016-02-11 23:45:00',
+      bucket_count: '371',
+      earliest_record_timestamp: '2019-06-12 00:04:19',
+      latest_record_timestamp: '2019-07-12 23:45:36',
+      input_record_count: '4,675',
+      latest_bucket_timestamp: '2019-07-12 22:00:00',
     };
   }
 
@@ -60,22 +74,22 @@ export default function ({ getService }: FtrProviderContext) {
       job_id: expectedJobId,
       result_type: 'model_size_stats',
       model_bytes_exceeded: '0.0 B',
-      total_by_field_count: '59',
-      total_over_field_count: '0',
-      total_partition_field_count: '58',
+      total_by_field_count: '25',
+      total_over_field_count: '92',
+      total_partition_field_count: '3',
       bucket_allocation_failures_count: '0',
       memory_status: 'ok',
-      timestamp: '2016-02-11 23:30:00',
+      timestamp: '2019-07-12 20:00:00',
     };
   }
 
   const calendarId = `wizard-test-calendar_${Date.now()}`;
 
-  describe('multi metric', function () {
+  describe('population', function () {
     this.tags(['ml']);
     before(async () => {
-      await esArchiver.loadIfNeeded('x-pack/test/functional/es_archives/ml/farequote');
-      await ml.testResources.createIndexPatternIfNeeded('ft_farequote', '@timestamp');
+      await esArchiver.loadIfNeeded('x-pack/test/functional/es_archives/ml/ecommerce');
+      await ml.testResources.createIndexPatternIfNeeded('ft_ecommerce', 'order_date');
       await ml.testResources.setKibanaTimeZoneToUTC();
 
       await ml.api.createCalendar(calendarId);
@@ -84,10 +98,10 @@ export default function ({ getService }: FtrProviderContext) {
 
     after(async () => {
       await ml.api.cleanMlIndices();
-      await ml.testResources.deleteIndexPatternByTitle('ft_farequote');
+      await ml.testResources.deleteIndexPatternByTitle('ft_ecommerce');
     });
 
-    it('job creation loads the multi metric wizard for the source data', async () => {
+    it('job creation loads the population wizard for the source data', async () => {
       await ml.testExecution.logTestStep('job creation loads the job management page');
       await ml.navigation.navigateToMl();
       await ml.navigation.navigateToJobManagement();
@@ -96,20 +110,20 @@ export default function ({ getService }: FtrProviderContext) {
       await ml.jobManagement.navigateToNewJobSourceSelection();
 
       await ml.testExecution.logTestStep('job creation loads the job type selection page');
-      await ml.jobSourceSelection.selectSourceForAnomalyDetectionJob('ft_farequote');
+      await ml.jobSourceSelection.selectSourceForAnomalyDetectionJob('ft_ecommerce');
 
-      await ml.testExecution.logTestStep('job creation loads the multi metric job wizard page');
-      await ml.jobTypeSelection.selectMultiMetricJob();
+      await ml.testExecution.logTestStep('job creation loads the population job wizard page');
+      await ml.jobTypeSelection.selectPopulationJob();
     });
 
-    it('job creation navigates through the multi metric wizard and sets all needed fields', async () => {
+    it('job creation navigates through the population wizard and sets all needed fields', async () => {
       await ml.testExecution.logTestStep('job creation displays the time range step');
       await ml.jobWizardCommon.assertTimeRangeSectionExists();
 
       await ml.testExecution.logTestStep('job creation sets the time range');
       await ml.jobWizardCommon.clickUseFullDataButton(
-        'Feb 7, 2016 @ 00:00:00.000',
-        'Feb 11, 2016 @ 23:59:54.000'
+        'Jun 12, 2019 @ 00:04:19.000',
+        'Jul 12, 2019 @ 23:45:36.000'
       );
 
       await ml.testExecution.logTestStep('job creation displays the event rate chart');
@@ -119,30 +133,42 @@ export default function ({ getService }: FtrProviderContext) {
       await ml.testExecution.logTestStep('job creation displays the pick fields step');
       await ml.jobWizardCommon.advanceToPickFieldsSection();
 
+      await ml.testExecution.logTestStep('job creation selects the population field');
+      await ml.jobWizardPopulation.assertPopulationFieldInputExists();
+      await ml.jobWizardPopulation.selectPopulationField(populationField);
+
       await ml.testExecution.logTestStep(
         'job creation selects detectors and displays detector previews'
       );
-      for (const [index, aggAndFieldIdentifier] of aggAndFieldIdentifiers.entries()) {
+      for (const [index, detector] of detectors.entries()) {
         await ml.jobWizardCommon.assertAggAndFieldInputExists();
-        await ml.jobWizardCommon.selectAggAndField(aggAndFieldIdentifier, false);
-        await ml.jobWizardCommon.assertDetectorPreviewExists(aggAndFieldIdentifier, index, 'LINE');
+        await ml.jobWizardCommon.selectAggAndField(detector.identifier, false);
+        await ml.jobWizardCommon.assertDetectorPreviewExists(detector.identifier, index, 'SCATTER');
       }
 
       await ml.testExecution.logTestStep(
-        'job creation inputs the split field and displays split cards'
+        'job creation inputs detector split fields and displays split cards'
       );
-      await ml.jobWizardMultiMetric.assertSplitFieldInputExists();
-      await ml.jobWizardMultiMetric.selectSplitField(splitField);
+      for (const [index, detector] of detectors.entries()) {
+        await ml.jobWizardPopulation.assertDetectorSplitFieldInputExists(index);
+        await ml.jobWizardPopulation.selectDetectorSplitField(index, detector.splitField);
 
-      await ml.jobWizardMultiMetric.assertDetectorSplitExists(splitField);
-      await ml.jobWizardMultiMetric.assertDetectorSplitFrontCardTitle('AAL');
-      await ml.jobWizardMultiMetric.assertDetectorSplitNumberOfBackCards(9);
-
-      await ml.jobWizardCommon.assertInfluencerSelection([splitField]);
+        await ml.jobWizardPopulation.assertDetectorSplitExists(index);
+        await ml.jobWizardPopulation.assertDetectorSplitFrontCardTitle(
+          index,
+          detector.frontCardTitle
+        );
+        await ml.jobWizardPopulation.assertDetectorSplitNumberOfBackCards(
+          index,
+          detector.numberOfBackCards
+        );
+      }
 
       await ml.testExecution.logTestStep('job creation displays the influencer field');
       await ml.jobWizardCommon.assertInfluencerInputExists();
-      await ml.jobWizardCommon.assertInfluencerSelection([splitField]);
+      await ml.jobWizardCommon.assertInfluencerSelection(
+        [populationField].concat(detectors.map((detector) => detector.splitField))
+      );
 
       await ml.testExecution.logTestStep('job creation inputs the bucket span');
       await ml.jobWizardCommon.assertBucketSpanInputExists();
@@ -219,27 +245,27 @@ export default function ({ getService }: FtrProviderContext) {
       );
 
       await ml.testExecution.logTestStep('job creation has detector results');
-      for (let i = 0; i < aggAndFieldIdentifiers.length; i++) {
+      for (let i = 0; i < detectors.length; i++) {
         await ml.api.assertDetectorResultsExist(jobId, i);
       }
     });
 
-    it('job cloning opens the existing job in the multi metric wizard', async () => {
+    it('job cloning opens the existing job in the population wizard', async () => {
       await ml.testExecution.logTestStep(
-        'job cloning clicks the clone action and loads the multi metric wizard'
+        'job cloning clicks the clone action and loads the population wizard'
       );
       await ml.jobTable.clickCloneJobAction(jobId);
-      await ml.jobTypeSelection.assertMultiMetricJobWizardOpen();
+      await ml.jobTypeSelection.assertPopulationJobWizardOpen();
     });
 
-    it('job cloning navigates through the multi metric wizard, checks and sets all needed fields', async () => {
+    it('job cloning navigates through the population wizard, checks and sets all needed fields', async () => {
       await ml.testExecution.logTestStep('job cloning displays the time range step');
       await ml.jobWizardCommon.assertTimeRangeSectionExists();
 
       await ml.testExecution.logTestStep('job cloning sets the time range');
       await ml.jobWizardCommon.clickUseFullDataButton(
-        'Feb 7, 2016 @ 00:00:00.000',
-        'Feb 11, 2016 @ 23:59:54.000'
+        'Jun 12, 2019 @ 00:04:19.000',
+        'Jul 12, 2019 @ 23:45:36.000'
       );
 
       await ml.testExecution.logTestStep('job cloning displays the event rate chart');
@@ -249,24 +275,35 @@ export default function ({ getService }: FtrProviderContext) {
       await ml.testExecution.logTestStep('job cloning displays the pick fields step');
       await ml.jobWizardCommon.advanceToPickFieldsSection();
 
+      await ml.testExecution.logTestStep('job cloning pre-fills the population field');
+      await ml.jobWizardPopulation.assertPopulationFieldInputExists();
+      await ml.jobWizardPopulation.assertPopulationFieldSelection([populationField]);
+
       await ml.testExecution.logTestStep(
         'job cloning pre-fills detectors and shows preview with split cards'
       );
-      for (const [index, aggAndFieldIdentifier] of aggAndFieldIdentifiers.entries()) {
-        await ml.jobWizardCommon.assertDetectorPreviewExists(aggAndFieldIdentifier, index, 'LINE');
+      for (const [index, detector] of detectors.entries()) {
+        await ml.jobWizardCommon.assertDetectorPreviewExists(detector.identifier, index, 'SCATTER');
+
+        await ml.jobWizardPopulation.assertDetectorSplitFieldSelection(index, [
+          detector.splitField,
+        ]);
+        await ml.jobWizardPopulation.assertDetectorSplitExists(index);
+        await ml.jobWizardPopulation.assertDetectorSplitFrontCardTitle(
+          index,
+          detector.frontCardTitle
+        );
+        await ml.jobWizardPopulation.assertDetectorSplitNumberOfBackCards(
+          index,
+          detector.numberOfBackCards
+        );
       }
-
-      await ml.jobWizardMultiMetric.assertDetectorSplitExists(splitField);
-      await ml.jobWizardMultiMetric.assertDetectorSplitFrontCardTitle('AAL');
-      await ml.jobWizardMultiMetric.assertDetectorSplitNumberOfBackCards(9);
-
-      await ml.testExecution.logTestStep('job cloning pre-fills the split field');
-      await ml.jobWizardMultiMetric.assertSplitFieldInputExists();
-      await ml.jobWizardMultiMetric.assertSplitFieldSelection([splitField]);
 
       await ml.testExecution.logTestStep('job cloning pre-fills influencers');
       await ml.jobWizardCommon.assertInfluencerInputExists();
-      await ml.jobWizardCommon.assertInfluencerSelection([splitField]);
+      await ml.jobWizardCommon.assertInfluencerSelection(
+        [populationField].concat(detectors.map((detector) => detector.splitField))
+      );
 
       await ml.testExecution.logTestStep('job cloning pre-fills the bucket span');
       await ml.jobWizardCommon.assertBucketSpanInputExists();
@@ -351,7 +388,7 @@ export default function ({ getService }: FtrProviderContext) {
       );
 
       await ml.testExecution.logTestStep('job cloning has detector results');
-      for (let i = 0; i < aggAndFieldIdentifiers.length; i++) {
+      for (let i = 0; i < detectors.length; i++) {
         await ml.api.assertDetectorResultsExist(jobId, i);
       }
     });
